refactor(af-move-type): internal reorg
